SearchEngine/requirements.txt
python-dotenv
streamlit
langchain
langchain-community
pypdf
duckduckgo-search
